test(kadena-cli): improve msw mocks
