+++ packages/tools/kadena-cli/src/commands/account/utils/__tests__/mocks.ts
@@ -1,20 +1,6 @@
 import type { ChainId } from '@kadena/client';
-import type { INetworkCreateOptions } from '../../../networks/utils/networkHelpers.js';
+import { testNetworkConfigMock } from '../../../../mocks/network.js';
 import type { Predicate } from '../../types.js';
-
-export const testNetworkConfigMock: INetworkCreateOptions = {
-  networkHost: 'https://api.testnet.chainweb.com',
-  networkExplorerUrl: 'https://explorer.chainweb.com/testnet04',
-  networkId: 'testnet04',
-  network: 'testnet',
-};
-
-export const devNetConfigMock: INetworkCreateOptions = {
-  networkHost: 'http://localhost:8080',
-  networkExplorerUrl: 'http://localhost:8080/explorer',
-  networkId: 'development',
-  network: 'devnet',
-};
 
 export const defaultConfigMock = {
   accountAlias: 'accountAlias',
